backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
alembic
greenlet
requests
pytest
httpx
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart